pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-name"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "pyserial~=3.5"
]

[tool.isort]
profile = "black"

